Fixes NuGet issue #1 and #4.
Earlier I claimed I fixed #1, but I guess I didn't test it enough. $(UseDebugLibraries) isn't defined till later is the MSBuild, so I was using something unset, at a time when I needed it to be set (for the <ItemsGroup> elements or whatever).
Now that I deferred the lookup of $(UseDebugLibraries) to a TailInject.props (read by the .targets file injected by NuGet into the tail of the .vcxproj file), it works. I also included backup functionality, in the event $(UseDebugLibraries) isn't defined (which may be the case of a MSVC project upgrade, or hand editing), and I have tested this and it works as I expect (in the case of a MSVC project upgrade, it explicity defines a debug RuntimeLibrary setting, in the case UseDebugLibraries was hand removed it isn't a debug configuration anymore).
Oh, also I figured out #4; I think I had to defer the WXDEBUGFLAG definition (see above #1), and also defer this <ItemsGroup> defining which dlls needed to be copied.